refactor distance module into new routing module
